+++ Assets/Tests/SocketLayer/AckSystem/AckSystemTest_Reliable.cs
@@ -11,7 +11,7 @@ namespace Mirage.SocketLayer.Tests.AckSystemTests
     {
         private class Time : ITime
         {
-            public float Now { get; set; }
+            public double Now { get; set; }
         }
 
         private class BadSocket
